refactor: create one packages for all smart contracts (#77)
