+++ packages/ethereum-storage/package.json
@@ -7,8 +7,7 @@
   "description": "Request Network storage layer based on ethereum.",
   "keywords": [
     "requestnetwork",
-    "ethereum-storage",
-    "smart-contracts"
+    "ethereum-storage"
   ],
   "repository": {
     "type": "git",
@@ -22,35 +21,26 @@
   "engines": {
     "node": ">=8.0.0"
   },
-  "main": "dist/src/lib/index.js",
-  "types": "dist/src/lib/index.d.ts",
+  "main": "dist/src/index.js",
+  "types": "dist/src/index.d.ts",
   "directories": {
-    "lib": "src/lib",
-    "test": "test/lib"
+    "lib": "src",
+    "test": "test"
   },
   "files": [
     "dist"
   ],
   "scripts": {
-    "build:lib": "tsc -b",
-    "build:sol": "truffle compile --contracts_directory=./src",
-    "build": "yarn run build:lib && yarn run build:sol",
-    "clean:lib": "shx rm -rf dist",
-    "clean:sol": "shx rm -rf build",
-    "clean": "yarn run clean:lib && yarn run clean:sol",
-    "lint:lib": "tslint --project . && eslint \"src/lib/**/*.ts\"",
-    "lint:sol": "solium --dir src/contracts",
-    "lint": "yarn run lint:lib && yarn run lint:sol",
+    "build": "tsc -b",
+    "clean": "shx rm -rf dist",
+    "lint": "tslint --project . && eslint \"src/**/*.ts\"",
     "lint-staged": "lint-staged",
-    "ganache": "ganache-cli -l 90000000 -p 8545 -m \"candy maple cake sugar pudding cream honey rich smooth crumble sweet treat\"",
-    "deploy": "truffle --contracts_directory=./src deploy",
-    "test:lib": "nyc mocha --timeout=5000 --extension ts --require ts-node/register --require source-map-support/register \"test/lib/**/*.ts\"",
-    "test:lib:watch": "yarn test --watch",
-    "test:sol": "truffle test --contracts_directory=./src test/contracts/*.js",
-    "test": "yarn run test:lib && yarn run test:sol",
+    "test": "nyc mocha --timeout=5000 --require ts-node/register --require source-map-support/register \"test/**/*.ts\"",
+    "test:watch": "nyc mocha --watch --watch-extensions ts --timeout=5000 --require ts-node/register --require source-map-support/register \"test/**/*.ts\"",
     "init-ipfs": "node scripts/init-ipfs.js"
   },
   "dependencies": {
+    "@requestnetwork/smart-contracts": "0.1.0",
     "@requestnetwork/types": "0.9.1",
     "@requestnetwork/utils": "0.8.0",
     "bluebird": "3.5.3",
@@ -65,7 +55,6 @@
     "yargs": "12.0.5"
   },
   "devDependencies": {
-    "@openzeppelin/contracts": "2.4.0",
     "@truffle/hdwallet-provider": "1.0.18",
     "@types/bluebird": "3.5.27",
     "@types/chai": "4.1.7",
@@ -85,7 +74,6 @@
     "eslint-plugin-spellcheck": "0.0.11",
     "eslint-plugin-typescript": "0.14.0",
     "fetch-mock": "7.3.3",
-    "ganache-cli": "6.3.0",
     "lint-staged": "8.1.3",
     "mocha": "6.2.2",
     "nyc": "13.2.0",
@@ -96,7 +84,6 @@
     "sinon": "7.5.0",
     "solium": "1.2.2",
     "source-map-support": "0.5.13",
-    "truffle": "5.0.3",
     "ts-node": "8.5.4",
     "tslint": "5.12.1",
     "typescript": "3.7.2",
