+++ packages/ethereum-storage/src/ethereum-utils.ts
