+++ packages/ethereum-storage/tsconfig.json
@@ -5,12 +5,6 @@
     "rootDir": ".",
     "resolveJsonModule": true
   },
-  "include": ["./src/lib/**/*", "./artifacts/**/*"],
-  "references": [{ "path": "../types" }, { "path": "../utils" }],
-  "files": [
-    "./artifacts/RequestHashStorage/0.1.0.json",
-    "./artifacts/RequestHashStorage/artifacts.json",
-    "./artifacts/RequestHashSubmitter/RequestOpenHashSubmitter/0.1.0.json",
-    "./artifacts/RequestHashSubmitter/RequestOpenHashSubmitter/artifacts.json"
-  ]
+  "include": ["./src/**/*"],
+  "references": [{ "path": "../types" }, { "path": "../utils" }, { "path": "../smart-contracts" }]
 }
