+++ packages/ethereum-storage/tslint.json
@@ -1,6 +1,3 @@
 {
-  "extends": "../../tslint.json",
-  "linterOptions": {
-    "exclude": ["artifacts/**/*"]
-  }
+  "extends": "../../tslint.json"
 }
