+++ packages/smart-contracts/.gitignore
