+++ packages/smart-contracts/artifacts/RequestHashSubmitter/RequestOpenHashSubmitter/artifacts.json
