+++ packages/smart-contracts/src/lib/index.ts
@@ -1,0 +1,7 @@
+import * as requestHashStorageArtifact from './requestHashStorageArtifact';
+import * as requestHashSubmitterArtifact from './requestHashSubmitterArtifact';
+
+export {
+  requestHashStorageArtifact,
+  requestHashSubmitterArtifact,
+};
