+++ packages/smart-contracts/src/lib/requestHashStorageArtifact.ts
