+++ packages/smart-contracts/src/lib/requestHashSubmitterArtifact.ts
