+++ packages/smart-contracts/test/contracts/bytes-test.js
