+++ packages/smart-contracts/tslint.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../../tslint.json",
+  "linterOptions": {
+    "exclude": ["artifacts/**/*"]
+  }
+}
